--- app/src/main/java/com/techcourse/service/BusinessMethod.java
@@ -1,7 +1,0 @@
-package com.techcourse.service;
-
-@FunctionalInterface
-public interface BusinessMethod<T, R> {
-
-    R method(T t) throws Exception;
-}
