+++ jdbc/src/main/java/com/interface21/transaction/support/TransactionalMethod.java
@@ -1,0 +1,9 @@
+package com.interface21.transaction.support;
+
+import java.sql.Connection;
+
+@FunctionalInterface
+public interface TransactionalMethod<R> {
+
+    R method(Connection connection) throws Exception;
+}
